# backend/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Aura API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/mediaserver/library/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve one filtered, sorted page from the warmed media-server library cache.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MediaServer"
                ],
                "summary": "Get Library Items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Library Title (repeat for multiple)",
                        "name": "library_titles",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Title search",
                        "name": "search_title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Library search",
                        "name": "search_library",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exact TMDB ID search",
                        "name": "search_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Release year",
                        "name": "search_year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Database filter (inDB or notInDB)",
                        "name": "filter_in_db",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ignore status or mode",
                        "name": "filter_ignored",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "MediUX set availability",
                        "name": "filter_has_sets",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort option",
                        "name": "sort_option",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc or desc)",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (maximum 1000)",
                        "name": "items_per_page",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Refresh server cache before reading",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/routes_ms.GetLibrarySectionItems_Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized (only when Auth.Enabled=true)",
                        "schema": {
                            "$ref": "#/definitions/httpx.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResponse"
                        }
                    }
                }
            }
        },
...
        "/api/mediaserver/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refresh the metadata of a media item on the media server. This endpoint accepts a rating_key for the media item to be refreshed and an optional refresh_rating_key to specify which metadata entry to refresh. It triggers a metadata refresh on the media server for the specified media item, allowing clients to update the displayed information for that item after changes have been made on the media server or to fix any metadata issues.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MediaServer"
                ],
                "summary": "Refresh Media Item Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rating Key of the media item to refresh",
                        "name": "rating_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Rating Key to specify which metadata entry to refresh",
                        "name": "refresh_rating_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpx.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/routes_ms.RefreshMediaItemMetadata_Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized (only when Auth.Enabled=true)",
                        "schema": {
                            "$ref": "#/definitions/httpx.UnauthorizedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/logging.LogErrorInfo"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "httpx.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid or expired token"
                }
            }
        },
...
        "logging.LogErrorInfo": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "object"
                },
                "function": {
                    "type": "string",
                    "example": "Reload"
                },
                "help": {
                    "type": "string",
                    "example": "Check server logs for more details"
                },
                "line_number": {
                    "type": "integer",
                    "example": 42
                },
                "message": {
                    "type": "string",
                    "example": "Invalid or expired token"
                }
            }
        },
...
        "models.DBSavedSet": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "selected_types": {
                    "$ref": "#/definitions/models.SelectedTypes"
                },
                "user_created": {
                    "type": "string"
                }
            }
        },
...
        "models.MediaItem": {
            "type": "object",
            "properties": {
                "added_at": {
                    "description": "Added timestamp",
                    "type": "integer"
                },
                "content_rating": {
                    "description": "Used in MediaItem Details Page - For more information about the item",
                    "type": "string"
                },
                "db_saved_sets": {
                    "description": "Used in MediaItem Details Page - For poster sets",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DBSavedSet"
                    }
                },
                "guids": {
                    "description": "Used in MediaItem Details Page - For ratings",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItemGuid"
                    }
                },
                "has_mediux_sets": {
                    "description": "Used in Home Page - sorting and filtering",
                    "type": "boolean"
                },
                "ignored_in_db": {
                    "description": "Whether the item is marked as ignored",
                    "type": "boolean"
                },
                "ignored_mode": {
                    "description": "Mode of ignoring (e.g., \"always\", \"until-set-available\", \"until-new-set-available\")",
                    "type": "string"
                },
                "ignored_sets": {
                    "description": "List of set IDs that were present when the item was ignored (used for \"until-new-set-available\" mode)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latest_episode_added_at": {
                    "type": "integer"
                },
                "library_title": {
                    "description": "Title of the library/section the item belongs to",
                    "type": "string"
                },
                "movie": {
                    "description": "Present if Type is \"movie\"; Contains file info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.MediaItemMovie"
                        }
                    ]
                },
                "rating_key": {
                    "description": "RatingKey is the internal ID from the media server",
                    "type": "string"
                },
                "released_at": {
                    "description": "Release date timestamp",
                    "type": "integer"
                },
                "series": {
                    "description": "Present if Type is \"show\"; Contains seasons and episodes info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.MediaItemSeries"
                        }
                    ]
                },
                "summary": {
                    "description": "Summary or description of the media item",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the media item",
                    "type": "string"
                },
                "tmdb_id": {
                    "description": "TMDB ID of the media item",
                    "type": "string"
                },
                "type": {
                    "description": "\"movie\" or \"show\"",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Last updated timestamp",
                    "type": "integer"
                },
                "year": {
                    "description": "Release year of the media item",
                    "type": "integer"
                }
            }
        },
        "models.MediaItemEpisode": {
            "type": "object",
            "properties": {
                "added_at": {
                    "type": "integer"
                },
                "episode_number": {
                    "type": "integer"
                },
                "file": {
                    "$ref": "#/definitions/models.MediaItemFile"
                },
                "rating_key": {
                    "type": "string"
                },
                "season_number": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.MediaItemFile": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "models.MediaItemGuid": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                }
            }
        },
        "models.MediaItemMovie": {
            "type": "object",
            "properties": {
                "file": {
                    "$ref": "#/definitions/models.MediaItemFile"
                }
            }
        },
        "models.MediaItemSeason": {
            "type": "object",
            "properties": {
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItemEpisode"
                    }
                },
                "rating_key": {
                    "type": "string"
                },
                "season_number": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.MediaItemSeries": {
            "type": "object",
            "properties": {
                "episode_count": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "season_count": {
                    "type": "integer"
                },
                "seasons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItemSeason"
                    }
                }
            }
        },
...
        "models.SelectedTypes": {
            "type": "object",
            "properties": {
                "backdrop": {
                    "type": "boolean"
                },
                "poster": {
                    "type": "boolean"
                },
                "season_poster": {
                    "type": "boolean"
                },
                "special_season_poster": {
                    "type": "boolean"
                },
                "titlecard": {
                    "type": "boolean"
                }
            }
        },
...
        "routes_ms.GetLibrarySectionItems_Response": {
            "type": "object",
            "properties": {
                "has_episode_added_at": {
                    "type": "boolean"
                },
                "has_updated_at": {
                    "type": "boolean"
                },
                "media_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaItem"
                    }
                },
                "total_items": {
                    "type": "integer"
                }
            }
        },
...
        "routes_ms.RefreshMediaItemMetadata_Response": {
            "type": "object",
            "properties": {
                "refreshed": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}